Moved files outside component
