+++ react/src/lib/components/ColorLegends/WebVizContinuousLegend.tsx
