+++ react/src/lib/components/ColorLegends/WebVizDiscreteLegend.tsx
